use `BigDecimal` for better number compatibility

+++ examples/CancelOrderByClientIdExample.java
@@ -29,7 +29,7 @@ public class CancelOrderByClientIdExample {
         }
 
         ParadexMarketSummaryDTO btcMarketSummary = markets.get(0);
-        double currentPrice = btcMarketSummary.getMarkPrice();
+        double currentPrice = btcMarketSummary.getMarkPrice().doubleValue();
 
         double orderPrice = BigDecimal.valueOf(currentPrice)  // -5% from current price
                 .multiply(BigDecimal.valueOf(0.95))
