+++ examples/ModifyOrderExample.java
@@ -26,7 +26,7 @@ public class ModifyOrderExample {
         }
 
         ParadexMarketSummaryDTO btcMarketSummary = markets.get(0);
-        double currentPrice = btcMarketSummary.getMarkPrice();
+        double currentPrice = btcMarketSummary.getMarkPrice().doubleValue();
 
         double orderPrice = BigDecimal.valueOf(currentPrice)  // -3% from current price
                 .multiply(BigDecimal.valueOf(0.97))
