+++ src/main/java/trade/paradex/api/dto/ParadexVaultBalanceDTO.java
@@ -3,6 +3,8 @@ package trade.paradex.api.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ParadexVaultBalanceDTO {
 
@@ -13,6 +15,6 @@ public class ParadexVaultBalanceDTO {
     private String token;
 
     @JsonProperty("size")
-    private double size;
+    private BigDecimal size;
 
 }
